@@ -18,8 +18,8 @@
     "build": "npm run lint && npm run build:commonjs && npm run build:es",
     "build:commonjs": "rimraf ./lib && tsc -p . --outDir ./lib --module commonjs",
     "build:es": "rimraf ./es && tsc -p . --outDir ./es --module es2015 --moduleResolution Node",
-    "build-demo": "rimraf ./docs && cross-env NODE_ENV=production webpack -p --config webpack.demo.config.js",
-    "run-demos": "webpack-dev-server --hot --config webpack.demo.config.js --progress",
+    "build-demo": "rimraf ./docs && cross-env NODE_ENV=production webpack --mode production --config webpack.demo.config.js",
+    "run-demos": "webpack-dev-server --mode development -c webpack.demo.config.js",
     "lint": "eslint --ext .js,.jsx,.ts,.tsx ./src ./demo",
     "prepublishOnly": "npm run build",
     "start": "npm run run-demos"
@@ -56,41 +56,40 @@
     "@types/classnames": "^2.2.10",
     "@types/jsonp": "^0.2.0",
     "@types/react": "^16.9.49",
-    "@types/react-dom": "^16.9.8",
+    "@types/react-dom": "^16.9.14",
+    "@types/webpack-dev-server": "^4.7.2",
     "@typescript-eslint/eslint-plugin": "^4.0.1",
     "@typescript-eslint/parser": "^4.0.1",
     "cross-env": "7.0.2",
-    "css-loader": "4.2.2",
+    "css-loader": "^4.2.2",
     "eslint": "7.8.1",
     "eslint-config-prettier": "^6.11.0",
-    "eslint-loader": "4.0.2",
+    "eslint-loader": "^4.0.2",
     "eslint-plugin-babel": "^5.3.1",
     "eslint-plugin-import": "2.22.0",
     "eslint-plugin-jsx-a11y": "6.3.1",
     "eslint-plugin-prettier": "^3.1.4",
     "eslint-plugin-react": "7.20.6",
     "eslint-plugin-react-hooks": "4.1.0",
-    "file-loader": "6.1.0",
-    "html-webpack-plugin": "4.4.1",
+    "file-loader": "^6.1.0",
+    "html-webpack-plugin": "^5.5.0",
     "prettier": "^2.1.1",
-    "react": "16.8.0",
-    "react-dom": "16.8.0",
-    "react-hot-loader": "4.12.21",
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0",
     "rimraf": "^3.0.2",
-    "style-loader": "1.2.1",
-    "terser-webpack-plugin": "4.1.0",
-    "ts-loader": "^8.0.3",
+    "style-loader": "^1.2.1",
+    "ts-loader": "^9.2.8",
     "typescript": "4.0.2",
-    "url-loader": "4.1.0",
-    "webpack": "4.44.1",
-    "webpack-cli": "3.3.12",
-    "webpack-dev-server": "3.11.0"
+    "url-loader": "^4.1.0",
+    "webpack": "^5.72.0",
+    "webpack-cli": "^4.9.2",
+    "webpack-dev-server": "^4.8.1"
   },
   "dependencies": {
     "classnames": "^2.2.5",
     "jsonp": "^0.2.1"
   },
   "peerDependencies": {
-    "react": "^16.3.0 || ^17"
+    "react": "^16.3.0 || ^17 || ^18"
   }
 }
