@@ -1,6 +1,7 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
-import Root from './Root';
+import { createRoot } from 'react-dom/client';
+
+import Demo from './Demo';
 
 const rootElement = document.createElement('div');
 
@@ -8,4 +9,5 @@ if (!document.querySelector('div')) {
   document.body.appendChild(rootElement);
 }
 
-ReactDOM.render(<Root />, rootElement);
+const root = createRoot(rootElement);
+root.render(<Demo />);
